backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
qdrant-client==1.7.0
sentence-transformers==2.3.1
langchain==0.1.1
langchain-community==0.0.13
langgraph==0.0.20
google-generativeai==0.3.2
httpx==0.26.0
pypdf==3.17.4
python-docx==1.1.0
reportlab==4.0.9
tenacity==8.2.3
